プランをクラス毎ではなくJSONで管理
