+++ challengeA/yoshiharu_chiba/lib/json_plan.json
@@ -1,0 +1,17 @@
+[
+  {
+    "provider_name": "東京電力エナジーパートナー",
+    "plan_name": "従量電灯Bプラン",
+    "price": ""
+  },
+  {
+    "provider_name": "Looopでんき",
+    "plan_name": "おうちプラン",
+    "price": ""
+  },
+  {
+    "provider_name": "東京ガス",
+    "plan_name": "ずっとも電気１プラン",
+    "price": ""
+  }
+]
